Partial implementation of replicated GroupAdmin object
I've moved the default subgroup allocation code from subgroup_functions
to the new GroupAdmin object, and changed the way Group and ViewManager
are constructed to use this object instead of the old SubgroupInfo and
collection-of-functors to compute subgroup membership. This doesn't
compile or work yet, though; this commit is just to save state so I can
switch branches and change the old SubgroupInfo interface for the Azure
release.